+++ src/Yaref92.Events/Abstractions/IEventAggregator.cs
@@ -36,10 +36,16 @@ public interface IEventAggregator
     /// <param name="subscriber">The subscriber instance.</param>
     void SubscribeToEventType<T>(IEventSubscriber<T> subscriber) where T : class, IDomainEvent;
 
+    void SubscribeToEventType<T>(IAsyncEventSubscriber<T> subscriber) where T : class, IDomainEvent;
+
     /// <summary>
     /// Unsubscribes a synchronous subscriber from an event type.
     /// </summary>
     /// <typeparam name="T">The event type.</typeparam>
     /// <param name="subscriber">The subscriber instance.</param>
     void UnsubscribeFromEventType<T>(IEventSubscriber<T> subscriber) where T : class, IDomainEvent;
+
+    void UnsubscribeFromEventType<T>(IAsyncEventSubscriber<T> subscriber) where T : class, IDomainEvent;
+
+    Task PublishEventAsync<T>(T domainEvent, CancellationToken cancellationToken = default) where T : class, IDomainEvent;
 }